INST 346 Basketball Simulation Project
With my final group project in the 2022 INST 346 class that I took at the University of Maryland, my group and I created a simulation of a sample basketball league. This is based off of the NBA2k My League Simulation mode, and it is entirely automated through Python. The teams have players with overall ratings and the teams with better overall ratings are more likely to win.